backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chromadb==0.4.17
sentence-transformers==2.2.2
groq==0.4.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0